lib/chatbot-api/functions/GCC-clean/requirements.txt
pdf2image
Pillow
python-magic